Refactorisation EditProfileUi.cs : extraction du picker d'icône de profil en composant `ProfileIconPicker`.
